requirements.txt
asgiref==3.6.0
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
